Move _copy_visual/_rebuild_mesh to _utils; fix docstring language
