+++ src/cedalion/geometry/photogrammetry/anonymization/_utils.py
@@ -1,13 +1,14 @@
-"""Private helpers shared by ``preprocessing`` and ``mask``.
+"""Shared helpers used by ``preprocessing`` and ``mask``.
 
-Centralizes patterns that would otherwise drift across the pipeline:
-texture-image lookup, face/vertex reindexing, affine application to plain
-arrays and to ``LabeledPoints``, ear midpoint, and the upper-head centroid.
+Consolidates texture-image lookup, mesh rebuilding, face/vertex reindexing,
+affine application, landmark transforms, ear midpoint, and upper-head centroid
+to avoid duplicating these across sibling modules.
 """
 
 from typing import Callable
 
 import numpy as np
+import trimesh
 
 import cedalion.typing as cdt
 
@@ -91,3 +92,60 @@ def _upper_head_centroid(
     upper = vertices[vertices[:, 0] > x_min + 0.6 * (x_max - x_min)]
     centroid = upper.mean(axis=0)
     return centroid, x_max
+
+
+def _copy_visual(src_mesh, dst_mesh, vertex_index=None) -> None:
+    """Copy a trimesh ``visual`` onto ``dst_mesh``, optionally reindexing.
+
+    Passing ``visual=old.visual`` to a new ``Trimesh(...)`` silently downgrades
+    a ``TextureVisuals`` to ``ColorVisuals`` because the visual still
+    back-references the source mesh's vertex count. Rebuilding it from scratch
+    preserves TextureVisuals.
+
+    Args:
+        src_mesh: Source trimesh.
+        dst_mesh: Destination trimesh (mutated in place).
+        vertex_index: Indices into the source vertex array. If given, UVs /
+            vertex_colors are sliced by this index. Leave None when the
+            vertex count is unchanged.
+    """
+    src_visual = src_mesh.visual
+    uv = getattr(src_visual, "uv", None)
+    n_src = len(src_mesh.vertices)
+
+    if uv is not None and len(uv) == n_src:
+        uv_arr = np.asarray(uv)
+        if vertex_index is not None:
+            uv_arr = uv_arr[vertex_index]
+        image = _resolve_texture_image(src_visual)
+        material = getattr(src_visual, "material", None)
+        dst_mesh.visual = trimesh.visual.TextureVisuals(
+            uv=uv_arr,
+            image=image,
+            material=material,
+        )
+        return
+
+    vcol = getattr(src_visual, "vertex_colors", None)
+    if vcol is not None and len(vcol) == n_src:
+        vcol_arr = np.asarray(vcol)
+        if vertex_index is not None:
+            vcol_arr = vcol_arr[vertex_index]
+        dst_mesh.visual.vertex_colors = vcol_arr
+
+
+def _rebuild_mesh(
+    src_mesh: trimesh.Trimesh,
+    *,
+    vertices: np.ndarray,
+    faces: np.ndarray,
+    vertex_index: np.ndarray | None = None,
+) -> trimesh.Trimesh:
+    """Build a new ``Trimesh`` and re-attach the source visual.
+
+    Used by every transform in the pipeline to keep the
+    ``Trimesh(process=False)`` + ``_copy_visual`` pattern in one place.
+    """
+    new_mesh = trimesh.Trimesh(vertices=vertices, faces=faces, process=False)
+    _copy_visual(src_mesh, new_mesh, vertex_index=vertex_index)
+    return new_mesh
